admin Dashboard and frontend

+++ app/Http/Controllers/BackendController/UserController.php
@@ -5,33 +5,64 @@ namespace App\Http\Controllers\BackendController;
 use App\Http\Controllers\Controller;
 use App\Models\Backend\Admin\UserManagement\UserRole;
 use App\Models\User;
-use Carbon\Carbon;
 use Illuminate\Http\Request;
+use Carbon\Carbon;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
+use Intervention\Image\Facades\Image;
 
 class UserController extends Controller
 {
-    public function index()
+    public function index(Request $request)
     {
-        $user = User::with('user_role')->get();
+        $user = User::active()->with('user_role')->paginate(10);
         return view('backend.pages.users.index', compact('user'));
     }
+
+    public function search(Request $request){
+        $search = $request->input('search');
+
+        $user = User::when($search, function ($query, $search) {
+            return $query->where('name', 'like', '%' . $search . '%');
+        })->paginate(2); 
+    
+        return view('backend.pages.users.index', compact('user'));
+       
+    }
+    // public function search(Request $request){
+    //     $user = User::where('name', 'LIKE', '%'. $request->search_string .'%')
+    //     ->orWhere('email', 'LIKE', '%'. $request->search_string .'%')
+    //     ->orWhere('phone_number', 'LIKE', '%'. $request->search_string .'%')
+    //     ->orWhere('id','desc')
+    //     ->paginate(2);
+    //     if ($user->count() >=1) {
+    //         return view('backend.pages.users.index', compact('user'))->render();
+    //     }
+    //     else{
+    //         return response()->json([
+    //             'status'=>'nothing_found'
+    //         ]);
+    //     }
+    // }
     public function create()
     {
-        $user_roles = UserRole::orderBy('role_serial','DESC')->get();
+        $user_roles = UserRole::orderBy('role_serial', 'DESC')->get();
         return view('backend.pages.users.create', compact('user_roles'));
     }
     public function store(Request $request)
     {
-        // dd($request->all());
+
         $this->validate($request, [
             'name' => 'required',
             'role_serial' => 'required',
             'email' => 'required|unique:user_roles,role_serial',
             'phone' => 'required',
-            'password' => 'required', 'string', 'min:8', 'confirmed',
-            'image' => 'required','image'
+            'password' => 'required',
+            'string',
+            'min:8',
+            'confirmed',
+            'image' => 'required',
+            'image'
         ]);
 
 
@@ -44,10 +75,14 @@ class UserController extends Controller
         $user->created_at = Carbon::now()->toDateString();
         $user->creator = Auth::user()->id;
 
+
         if ($request->hasFile('image')) {
             $file = $request->file('image');
-            $fileName = time() . '_' . $file->getClientOriginalName(); 
+            $fileName = time() . '_' . $file->getClientOriginalName();
             $file->move(public_path('upload/user'), $fileName);
+            $image = Image::make(public_path('upload/user/' . $fileName));
+            $image->resize(300, 200);
+            $image->save(public_path('upload/user/' . $fileName));
             $user->image = 'upload/user/' . $fileName;
         }
         $user->save();
@@ -56,7 +91,7 @@ class UserController extends Controller
 
     public function edit($id)
     {
-        $user_roles = UserRole::orderBy('role_serial','DESC')->get();
+        $user_roles = UserRole::orderBy('role_serial', 'DESC')->get();
         $user = User::find($id);
         return view('backend.pages.users.edit', compact('user_roles', 'user'));
     }
@@ -68,7 +103,8 @@ class UserController extends Controller
             'name' => 'required',
             'role_serial' => 'required',
             'phone' => 'required',
-            'image' => 'required','image'
+            'image' => 'required',
+            'image'
         ]);
 
 
@@ -76,7 +112,7 @@ class UserController extends Controller
 
         if ($user->email != $request->email) {
             $this->validate($request, [
-                 'email' => ['required','unique:users,email,'.$user->id],
+                'email' => ['required', 'unique:users,email,' . $user->id],
             ]);
             $user->email = $request->email;
         }
@@ -86,16 +122,29 @@ class UserController extends Controller
         $user->created_at = Carbon::now()->toDateString();
         $user->creator = Auth::user()->id;
         $user->update();
- 
 
-        if($request->hasFile('image')){
-            if (file_exists(public_path($user->image))) {
+
+        if ($request->hasFile('image')) {
+
+            if (!empty($user->image) && file_exists(public_path($user->image))) {
                 unlink(public_path($user->image));
             }
+
+
             $file = $request->file('image');
-            $fileName = time() . '_' . $file->getClientOriginalName(); 
-            $file->move(public_path('upload/user'), $fileName);
+            $fileName = time() . '_' . $file->getClientOriginalName();
+
+
+            $image = Image::make($file)
+                ->fit(100, 100, function ($constraint) {
+                    $constraint->aspectRatio();
+                    $constraint->upsize();
+                })
+                ->save(public_path('upload/user/' . $fileName), 100);
+
+
             $user->image = 'upload/user/' . $fileName;
+
 
             $user->update();
         }
